im-util: Move filter types into a separate module
